dash/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
sqlalchemy==1.4.23
python-dotenv==0.19.0
tweepy==4.12.0
python-jose==3.3.0
passlib==1.7.4